refactor: 改进service中的缓存清理逻辑

+++ service.sh
@@ -1,2 +1,8 @@
 #!/system/bin/sh
-rm -r /data/system/package_cache/*
+MODPATH="/data/adb/modules/installerx_revived"
+if [ ! -f "$MODPATH/inxed" ]; then
+    rm -r /data/system/package_cache/*
+else
+    # 若希望“每次更新都清理”，可以取消注释掉下面这一行
+    # rm "$MODPATH/inxed"
+fi
